backend/requirements.txt
Flask
flask-cors
numpy
tensorflow
gunicorn
Pillow
